chore: move components related to recommended schedule to RecommendedSchedule directory
