+++ web/src/components/RecommendedSchedule/ScheduleDeleteButton.tsx
